v1.072
JobsSST - Added override library for default values that will run only once. Added filtering for station types for a faction to consider the sector as contested. Production stations in sectors must have a least one operational production module to be counted.
God - Added check for Eco (other mods can diff in for the check as well) that gives a warning when not detected and maxmodule setting is gt 10. Made maxmodule an adjustable value.
Menus - Cleaned up debug messages. Added maxmodule slider and incorporated warning if setting is higher than recommended due to mod detection. Changed wording on god quotas menu to be easier to understand.